db: extract database utilities
